requirements.txt
Flask==2.3.3
numpy==1.23.5
Pillow==9.5.0
tensorflow-cpu==2.10.0
h5py==3.7.0
requests==2.31.0
gunicorn==21.2.0